pyproject.toml
[project]
name = "lession1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-core>=0.3.67",
    "langchain-deepseek>=0.1.3",
    "langgraph>=0.5.0",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.1",
]
